osa4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -33,11 +33,26 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  try {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-  } catch(e) {
-    response.status(400).json({ error: 'invalid id' })
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!request.token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid '})
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user._id.toString() === decodedToken.id) {
+    console.log('Operation starts')
+    console.log('initial blogs array: ', user.blogs)
+    const updatedBlogsArray = user.blogs.filter(b => b._id.toString() !== blog._id.toString())
+    console.log('updated array: ', updatedBlogsArray)
+    user.blogs = updatedBlogsArray
+    console.log('user: ', user)
+    await user.save()
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'blog can only be removed by user who created it'})
   }
 })
